Rename input.h -> input_axis.h
